continue lib api cleanup

+++ libfreeipmi/src/api/ipmi-ssif-driver-api.c
@@ -136,7 +136,7 @@ ipmi_ssif_cmd_api (ipmi_ctx_t ctx,
     ssize_t read_len;
     int32_t hdr_len, cmd_len;
     fiid_field_t *tmpl = NULL;
-    int8_t rv = -1;
+    int rv = -1;
 
     if ((hdr_len = fiid_template_len_bytes (tmpl_hdr_kcs)) < 0)
       {
